webapp/requirements_fixed.txt
# FastAPI and server - use latest compatible versions
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Git operations
GitPython>=3.1.41

# Data validation - use version compatible with existing installation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0

# Async file operations
aiofiles>=23.2.1

# Testing dependencies
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0

# Development tools
python-dotenv>=1.0.0

# Additional dependencies that might be needed
PyGithub>=2.1.1  # For GitHub integration
python-gitlab>=4.0.0  # For GitLab integration